Update index.html
Fixes the long overdue typo. 


Ode to the Typographical Error
" The typographical error is a slippery thing and sly;
You can hunt till you are dizzy, but it somehow will get by.
It shrinks down in a corner, and it never stirs or peeps –
That typographical error, too small for human eyes –
The remainder of the issue may be clean as clean can be,
But the typographical error is the only thing you see. "
— Anonymous